// src/types/types.ts
interface NavLinksInterface {
  label: string;
  path: string;
}

export default NavLinksInterface;

// upcoming movies types

export interface UpcomingMoviesInterface {
  adult: boolean;
  backdrop_path: string;
  genre_ids: number[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
}